Linear Regression with multiple features
Problem Statement:

Suppose you are the CEO of a
restaurant franchise and are considering different cities for opening a new
outlet. The chain already has trucks in various cities and you have data for
profits and populations from the cities. You would like to use this data to help you select which city to expand
to next.

Learning Outcome: Tried out gradient descent and normal equations to find the optimum value of theta